pyproject.toml
[tool.poetry]
name = "busgranada"
version = "0.1.0"
description = ""
authors = ["Daniel Ranchal <daniranchal@wealize.digital>"]

[tool.poetry.dependencies]
python = "^3.9"
pyTelegramBotAPI = "^4.4.0"
python-dotenv = "^0.19.2"
httpx = "^0.22.0"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
